Tap Draw System Finish and Bug Fixes
* Finished tap draw system.
* Implemented brush size into BrushSO.
* Binded plate speed to events.
* Modified tap draw camera.

+++ Assets/Scripts/Runtime/Draw/SOs/PaintingSO.cs
@@ -97,5 +97,5 @@ public struct Layer
     public Color32 Color;
     public UnityEvent OnSelected;
     public void Selected() => OnSelected?.Invoke();
-    public int Data1;
+    public float Data1;
 }
